+++ package.json
@@ -14,6 +14,9 @@
   "author": "Stamen Design",
   "license": "ISC",
   "dependencies": {
+    "@mapbox/mapbox-gl-style-spec": "^13.28.0",
+    "lodash.isequal": "^4.5.0",
+    "mapbox-gl-style-recurse": "github:stamen/mapbox-gl-style-recurse#v0.2.0",
     "yargs": "^17.2.1"
   },
   "devDependencies": {
