+++ src/lib/constants.js
@@ -1,0 +1,4 @@
+export const REMOVE = 'remove';
+export const ADD = 'add';
+export const UPDATE = 'update';
+export const MOVE = 'move';
